+++ src/main/kotlin/com/ravit/stash/controller/externals/task/request/TaskPeriodExternalRequest.kt
@@ -1,0 +1,8 @@
+package com.ravit.stash.controller.externals.task.request
+
+import java.time.LocalDate
+
+data class TaskPeriodExternalRequest(
+    val startedAt: LocalDate?,
+    val completedAt: LocalDate?,
+)