+++ src/main/kotlin/com/ravit/stash/controller/externals/task/response/TaskPeriodExternalResponse.kt
@@ -1,0 +1,8 @@
+package com.ravit.stash.controller.externals.task.response
+
+import java.time.LocalDate
+
+data class TaskPeriodExternalResponse(
+    val startedAt: LocalDate?,
+    val completedAt: LocalDate?,
+)